requirements.txt
python==3.11
streamlit
google-generativeai
python-dotenv
PyPDF2
python-dotenv
langchain_google_genai
Pillow
